common Helpers are moved to multikueue.go
